+++ notUsed/Weapon3.java
@@ -19,9 +19,9 @@ public class Weapon3 extends NormalWeapon{
             if (input==null) return;
             Vec2d pos=owner.getBulletPosition();
             Vec2d velocity=input.limit(speed*7.5);
-            double size=this.size*0.5;
-            double health=this.health*0.7;
-            double damage=this.damage*0.8;
+            float size=this.size*0.5;
+            float health=this.health*0.7;
+            float damage=this.damage*0.8;
             for(int i=0;i<15;i++) {
                 shootBullet(pos.add(velocity.multiply(i/10d)), velocity.multiply(i/7d+0.5).add(Util.randomVec().limit(0.003)), size, health, damage);
             }
