+++ notUsed/Weapon4.java
@@ -19,10 +19,10 @@ public class Weapon4 extends NormalWeapon{
             if (input==null) return;
             Vec2d pos=owner.getBulletPosition();
 
-            double size=this.size*0.5;
-            double health=this.health*0.7;
-            double damage=this.damage*0.8;
-            double rot=0;
+            float size=this.size*0.5;
+            float health=this.health*0.7;
+            float damage=this.damage*0.8;
+            float rot=0;
             for(int i=0;i<20;i++) {
                 rot+=18d;
                 Vec2d velocity=input.limit(speed*7.5).rotate(rot);
