+++ notUsed/Weapon5.java
@@ -21,9 +21,9 @@ public class Weapon5 extends NormalWeapon{
             if (input==null) return;
             Vec2d pos=owner.getBulletPosition();
             Vec2d velocity=input.limit(speed*6).add(Util.randomVec().limit(0.005));
-            double size=this.size*0.5;
-            double health=this.health*300;
-            double damage=this.damage*0.5;
+            float size=this.size*0.5;
+            float health=this.health*300;
+            float damage=this.damage*0.5;
 
             if(last==0){
                 last=1;
