+++ src/main/java/big/modules/entity/DamageSource.java
@@ -2,9 +2,9 @@ package big.modules.entity;
 
 public class DamageSource {
     public Long id;
-    public double damage;
+    public float damage;
     public long time;
-    public DamageSource(Long id, double damage) {
+    public DamageSource(Long id, float damage) {
         this.id = id;
         this.damage = damage;
         this.time = System.currentTimeMillis();
@@ -12,7 +12,7 @@ public class DamageSource {
     public boolean isExpired() {
         return System.currentTimeMillis() - time > 5000;
     }
-    public void increase(double dmg){
+    public void increase(float dmg){
         this.damage+=dmg;
         this.time=System.currentTimeMillis();
     }
